backend/requirements.txt
# OrderFlow Backend - Python Dependencies

# Core Framework
Flask==3.1.0
Flask-CORS==5.0.0
python-dotenv==1.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7

# Authentication & Security
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1

# Data Validation & Serialization
marshmallow==3.24.0
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0

# Development Tools
pytest==8.3.4
httpie==3.2.4
